# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.betaVerifyRequest:
    properties:
      passphrase:
        type: string
    required:
    - passphrase
    type: object
  handlers.cacheKeyStats:
    properties:
      key:
        type: string
      memory_bytes:
        type: integer
      ttl_seconds:
        type: integer
    type: object
  handlers.cacheStatsResponse:
    properties:
      key_count:
        type: integer
      keys:
        items:
          $ref: '#/definitions/handlers.cacheKeyStats'
        type: array
    type: object
...
  handlers.logoutRequest:
    properties:
      refresh_token:
        description: オプション（下位互換性のため）
        type: string
    type: object
...
  handlers.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  handlers.testLoginRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
    required:
    - display_name
    - email
    type: object
  handlers.triggerNotificationRequest:
    properties:
      type:
        type: string
    required:
    - type
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: RoambleのバックエンドAPIドキュメント
  title: Roamble API
  version: "1.0"
paths:
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: アクセストークンとリフレッシュトークンをブラックリストに登録して無効化する
      parameters:
      - description: リフレッシュトークン（オプション）
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.logoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: ログアウト
      tags:
      - Auth
...
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: リフレッシュトークンを使用して新しいアクセストークンを取得する
      parameters:
      - description: リフレッシュトークン
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: トークンリフレッシュ
      tags:
      - Auth
...
  /api/beta/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 合言葉
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.betaVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ベータ合言葉照合
      tags:
      - Beta
  /api/dev/auth/test-login:
    post:
      consumes:
      - application/json
      description: 指定emailのユーザーでテストログインし、JWTアクセストークン/リフレッシュトークンを返す（development環境のみ有効）
      parameters:
      - description: テストログイン情報
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.testLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: テストログイン（開発用）
      tags:
      - Dev
  /api/dev/notifications/trigger:
    post:
      consumes:
      - application/json
      description: 指定した通知タイプを即時実行する（development環境のみ有効）
      parameters:
      - description: 通知タイプ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.triggerNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 通知即時発火（開発用）
      tags:
      - Dev
  /api/dev/suggestions/cache:
    delete:
      description: 提案関連のRedisキャッシュを削除する（development環境のみ有効）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 提案キャッシュ削除（開発用）
      tags:
      - Dev
  /api/dev/suggestions/stats:
    get:
      description: 提案関連キャッシュのキー数・TTL・メモリ使用量を返す（development環境のみ有効）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.cacheStatsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 提案キャッシュ統計取得（開発用）
      tags:
      - Dev
...
securityDefinitions:
  BearerAuth:
    description: JWTアクセストークン（Bearer形式）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"